Rearrange settings
