+++ AndroidManifest.xml
@@ -94,7 +94,7 @@
         -->
         <activity
             android:name="com.android.launcher3.SettingsActivity"
-            android:label="@string/settings_button_text"
+            android:label="@string/settings_title"
             android:theme="@android:style/Theme.DeviceDefault.Settings"
             android:parentActivityName="com.android.launcher3.Launcher"
             android:autoRemoveFromRecents="true">
@@ -103,6 +103,38 @@
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
+
+        <!-- Homescreen -->
+        <activity
+            android:name="com.android.launcher3.Homescreen"
+            android:label="@string/home_category_title"
+            android:theme="@android:style/Theme.DeviceDefault.Settings"
+            android:autoRemoveFromRecents="true"
+            android:parentActivityName="com.android.launcher3.SettingsActivity" />
+
+        <!-- Icons -->
+        <activity
+            android:name="com.android.launcher3.Icons"
+            android:label="@string/icons_category_title"
+            android:theme="@android:style/Theme.DeviceDefault.Settings"
+            android:autoRemoveFromRecents="true"
+            android:parentActivityName="com.android.launcher3.SettingsActivity" />
+
+        <!-- App drawer -->
+        <activity
+            android:name="com.android.launcher3.AppDrawer"
+            android:label="@string/app_drawer_category_title"
+            android:theme="@android:style/Theme.DeviceDefault.Settings"
+            android:autoRemoveFromRecents="true"
+            android:parentActivityName="com.android.launcher3.SettingsActivity" />
+
+        <!-- Gestures -->
+        <activity
+            android:name="com.android.launcher3.Gestures"
+            android:label="@string/gestures_category_title"
+            android:theme="@android:style/Theme.DeviceDefault.Settings"
+            android:autoRemoveFromRecents="true"
+            android:parentActivityName="com.android.launcher3.SettingsActivity" />
 
         <!--
         The settings provider contains Home's data, like the workspace favorites. The permissions
