+++ res/values/du_strings.xml
@@ -17,26 +17,40 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
 
+    <!-- General -->
+    <string name="settings_title">Settings</string>
+
+    <!-- Categories -->
+    <string name="home_category_title">Home screen</string>
+    <string name="home_category_summary">Set Google now panel and more</string>
+    <string name="icons_category_title">Icons</string>
+    <string name="icons_category_summary">Set icon shape and more</string>
+    <string name="app_drawer_category_title">App drawer</string>
+    <string name="app_drawer_category_summary">Customize your app drawer</string>
+    <string name="gestures_category_title">Gestures</string>
+    <string name="gestures_category_summary">Enhance your experience</string>
+
+    <!-- Sub categories -->
+    <string name="home_grid_title">Grid layout</string>
+
     <!-- Gestures -->
-    <string name="allow_swipe_down_clearall_title">Clear all on swipe down from recents view</string>
-    <string name="allow_swipe_down_clearall_desc">Doesn\'t works if P navbar gestures are enable</string>
+    <string name="allow_swipe_down_clearall_title">Clear all from recents view</string>
+    <string name="allow_swipe_down_clearall_desc">Swipe vertically downward in recents view to clear all apps.\n\nThis doesn\'t works if P navbar gestures are enable</string>
 
     <!-- Google Feed -->
-    <string name="settings_feed">Enable Google feed integration</string>
+    <string name="google_feed_title">Google now panel</string>
+    <string name="google_feed_summary">Swipe to the right of the homescreen to access the Google now panel</string>
 
     <!-- Grid -->
     <string name="num_default">Default</string>
-    <string name="grid_category_title">Columns and rows</string>
-    <string name="number_of_columns_title">Grid and app drawer columns</string>
-    <string name="number_of_rows_title">Grid and app drawer rows</string>
-    <string name="number_of_hotseat_title">Homescreen quicklaunch icons</string>
+    <string name="number_of_columns_title">Number of columns</string>
+    <string name="number_of_rows_title">Number of rows</string>
+    <string name="number_of_hotseat_title">Number of dock icons</string>
     <string name="state_loading">Applying\u2026</string>
-    <string name="app_drawer_category_title">App drawer</string>
 
     <!-- Show labels on icon -->
-    <string name="desktop_show_label_title">Homescreen Apps Labels</string>
-    <string name="desktop_show_label_summary">Show labels below the icons on desktop</string>
-    <string name="allapps_show_label_title">AppDrawer Apps Labels</string>
-    <string name="allapps_show_label_summary">Show labels below the icons in app drawer</string>
+    <string name="app_show_label_title">App labels</string>
+    <string name="desktop_show_label_summary">Show the labels of apps on the homscreen</string>
+    <string name="allapps_show_label_summary">Show the labels of apps in the app drawer</string>
 
 </resources>
