+++ src/com/android/launcher3/Launcher.java
@@ -2467,7 +2467,7 @@ public class Launcher extends BaseDraggingActivity implements LauncherExterns,
 
     @Override
     public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
-        if (SettingsActivity.KEY_FEED_INTEGRATION.equals(key)) {
+        if (Homescreen.KEY_FEED_INTEGRATION.equals(key)) {
             if (mLauncherTab == null) {
                     return;
             }
