+++ src/com/android/launcher3/Utilities.java
@@ -16,6 +16,9 @@
 
 package com.android.launcher3;
 
+import android.app.AlarmManager;
+import android.app.PendingIntent;
+import android.app.ProgressDialog;
 import android.app.WallpaperManager;
 import android.content.ComponentName;
 import android.content.Context;
@@ -38,6 +41,7 @@ import android.os.DeadObjectException;
 import android.os.Handler;
 import android.os.Message;
 import android.os.PowerManager;
+import android.os.SystemClock;
 import android.os.TransactionTooLargeException;
 import android.text.Spannable;
 import android.text.SpannableString;
@@ -50,6 +54,7 @@ import android.util.TypedValue;
 import android.view.View;
 
 import com.android.launcher3.config.FeatureFlags;
+import com.android.launcher3.util.LooperExecutor;
 
 import java.io.ByteArrayOutputStream;
 import java.io.Closeable;
@@ -105,6 +110,8 @@ public final class Utilities {
 
     public static final int SINGLE_FRAME_MS = 16;
 
+    private static final long WAIT_BEFORE_RESTART = 250;
+
     /**
      * Indicates if the device has a debug build. Should only be used to store additional info or
      * add extra logging and not for changing the app behavior.
@@ -141,7 +148,7 @@ public final class Utilities {
     }
 
     public static boolean showAllAppsLabel(Context context) {
-         return getPrefs(context).getBoolean(ALLAPPS_SHOW_LABEL, true);
+        return getPrefs(context).getBoolean(ALLAPPS_SHOW_LABEL, true);
     }
 
     public static boolean isPropertyEnabled(String propertyName) {
@@ -633,6 +640,30 @@ public final class Utilities {
 
     static boolean hasFeedIntegration(Context context) {
         SharedPreferences prefs = getPrefs(context.getApplicationContext());
-        return prefs.getBoolean(SettingsActivity.KEY_FEED_INTEGRATION, true);
+        return prefs.getBoolean(Homescreen.KEY_FEED_INTEGRATION, true);
+    }
+
+    public static void restart(final Context context) {
+        ProgressDialog.show(context, null, context.getString(R.string.state_loading), true, false);
+        new LooperExecutor(LauncherModel.getWorkerLooper()).execute(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    Thread.sleep(WAIT_BEFORE_RESTART);
+                } catch (Exception e) {
+                }
+
+                Intent intent = new Intent(Intent.ACTION_MAIN)
+                        .addCategory(Intent.CATEGORY_HOME)
+                        .setPackage(context.getPackageName())
+                        .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+
+                PendingIntent pendingIntent = PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_CANCEL_CURRENT | PendingIntent.FLAG_ONE_SHOT);
+                AlarmManager alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
+                alarmManager.setExact(AlarmManager.ELAPSED_REALTIME, SystemClock.elapsedRealtime() + 50, pendingIntent);
+
+                android.os.Process.killProcess(android.os.Process.myPid());
+            }
+        });
     }
 }
